--- src/main/kotlin/icfp2019/gameboard/gameboard.kt
@@ -1,7 +1,0 @@
-package icfp2019.gameboard
-
-import icfp2019.Node
-
-class GameBoard(twoDimensionalRepresentation: Array<Array<Node>>) {
-
-}
