+++ src/main/kotlin/icfp2019/objects/gameboard/Gameboard.kt
@@ -1,0 +1,6 @@
+package icfp2019.objects.gameboard
+
+import icfp2019.Problem
+
+class GameBoard(twoDimensionalRepr: Problem) {
+}
