+++ src/main/kotlin/icfp2019/states/transaction/TransitionEngine.kt
@@ -1,0 +1,14 @@
+package icfp2019.states.transaction
+
+import com.google.common.collect.ImmutableMap
+import icfp2019.GameState
+import icfp2019.objects.robot.RobotId
+import icfp2019.states.actions.Actions
+
+class TransitionEngine {
+  // -> Return GameState
+  fun apply(currentGameState: GameState, robotIdMap: ImmutableMap<RobotId, Action>): Nothing {
+    // Implement meaningful shit here
+    TODO("Implement apply")
+  }
+}
