--- src/main/kotlin/icfp2019/transaction/TransitionEngine.kt
@@ -1,14 +1,0 @@
-package icfp2019.transaction
-
-import com.google.common.collect.ImmutableMap
-import icfp2019.Actions
-import icfp2019.GameState
-import icfp2019.RobotId
-
-class TransitionEngine {
-  // -> Return GameState
-  fun apply(currentGameState: GameState, robotIdMap: ImmutableMap<RobotId, Actions>): Nothing {
-    // Implement meaningful shit here
-    TODO("Implement apply")
-  }
-}
